[BBC-9126] Storybook: MainContent.mdx (#974)
* [BBC-9126] MainContent.mdx

* [BBC-9126] MainContent.mdx

* styling

* Props

* Update src/layout/content/Content.story.mdx
