+++ src/layout/content/index.tsx
@@ -1,35 +1,5 @@
-import styled from 'styled-components'
+import { MainContentProps } from './Content'
+import { BottomContent, Content, StyledMainContent } from './Content.style'
 
-import { responsiveBreakpoints } from '../../_utils/branding'
-import { MainContent } from './content'
-
-export const Content = styled.div.attrs(() => ({
-  role: 'presentation',
-}))``
-
-export const BottomContent = styled.div.attrs(() => ({
-  role: 'presentation',
-}))``
-
-const StyledMainContent = styled(MainContent)`
-  & {
-    display: flex;
-    flex-direction: column;
-    min-height: calc(100vh - ${props => props.topBarHeight});
-    min-height: stretch;
-    justify-content: space-between;
-  }
-
-  & > *:not(${Content}, ${BottomContent}) {
-    flex-grow: 1;
-  }
-
-  @media (${responsiveBreakpoints.isMediaLarge}) {
-    & {
-      justify-content: flex-start;
-    }
-  }
-`
-
-export { StyledMainContent as MainContent }
+export { StyledMainContent as MainContent, BottomContent, Content, MainContentProps }
 export default StyledMainContent
